fix(config): reject non-positive thresholds, stop fractional ones truncating
Six review findings; all were real.

The worst was a log flood. `slow-query.threshold: 0` was accepted and matched
every successful query, so the middleware reported `slow-query` on all of them
— into the log sink redaction exists to protect. `0.5` did the same by a
different route: Settings.Duration converted a float with
`time.Duration(v) * time.Millisecond`, and time.Duration(0.5) truncates to 0,
so a sub-millisecond threshold became no threshold. The conversion scales
before converting now, and both duration settings reject non-positive values,
the way n-plus-one.threshold already did. A zero `n-plus-one.window` passed
the paired-key check and then left detection off, which is the same class.

checkOnlySelectsSomething inspected the resolved set, so it could not see the
typo it was written for: unknown names are dropped, `only: [selct-star]`
resolves to an empty map, and an empty whitelist is not a whitelist — every
rule runs, the opposite of the narrowing that was asked for. It takes the
configured list now and reports that case separately.

seq-scan's row-count escalation assigned SeverityWarning outright, undoing the
point of reading the registry default one line above: Register can replace a
built-in by name, so a seq-scan registered at CRITICAL would have reported the
wide scan as less severe than the narrow one. It only escalates upward now.

The migration snippet for the removed SlowQueryThreshold() did not compile —
it chained a field selector off Profile(), which returns two values. Replaced
with one that does, checked by building it against the package.

+++ analyzer/registry.go
@@ -72,7 +72,9 @@ func (s Settings) Duration(key string, def time.Duration) time.Duration {
 	case int64:
 		return time.Duration(v) * time.Millisecond
 	case float64:
-		return time.Duration(v) * time.Millisecond
+		// Scale before converting: time.Duration(0.5) truncates to 0, which
+		// turned a fractional-millisecond threshold into "no threshold".
+		return time.Duration(v * float64(time.Millisecond))
 	}
 	return def
 }
